POSHI-613 Move report into PoshiReportGenerator
