+++ modules/test/poshi/poshi-core/src/main/java/com/liferay/poshi/core/PoshiContext.java
@@ -28,7 +28,6 @@ import com.liferay.poshi.core.util.Validator;
 
 import java.io.File;
 import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
 
 import java.lang.reflect.Method;
@@ -39,13 +38,10 @@ import java.net.URL;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 
-import java.text.SimpleDateFormat;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Date;
 import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -62,7 +58,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang3.ArrayUtils;
 
 import org.dom4j.Attribute;
 import org.dom4j.Element;
@@ -255,6 +250,10 @@ public class PoshiContext {
 		String className, String namespace) {
 
 		return _rootElements.get("macro#" + namespace + "." + className);
+	}
+
+	public static Pattern getNamespaceClassCommandNamePattern() {
+		return _namespaceClassCommandNamePattern;
 	}
 
 	public static Properties getNamespacedClassCommandNameProperties(
@@ -555,6 +554,10 @@ public class PoshiContext {
 		return _testCaseDescriptions.get(classCommandName);
 	}
 
+	public static List<String> getTestCaseNamespacedClassCommandNamesNames() {
+		return _testCaseNamespacedClassCommandNames;
+	}
+
 	public static Element getTestCaseRootElement(
 		String className, String namespace) {
 
@@ -593,7 +596,6 @@ public class PoshiContext {
 		PoshiValidation.validate();
 
 		_writeTestCaseMethodNamesProperties();
-		_writeTestCSVReportFile();
 		_writeTestGeneratedProperties();
 	}
 
@@ -1867,93 +1869,6 @@ public class PoshiContext {
 		}
 
 		FileUtil.write("test.case.method.names.properties", sb.toString());
-	}
-
-	private static void _writeTestCSVReportFile() throws Exception {
-		PoshiProperties poshiProperties = PoshiProperties.getPoshiProperties();
-
-		if (poshiProperties.testCSVReportPropertyNames == null) {
-			return;
-		}
-
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("MM-dd-yyyy");
-
-		File reportCSVFile = new File(
-			StringUtil.combine(
-				"Report_", simpleDateFormat.format(new Date()), ".csv"));
-
-		try (FileWriter reportCSVFileWriter = new FileWriter(reportCSVFile)) {
-			List<String> reportLineItems = new ArrayList<>();
-
-			reportLineItems.add("Namespace");
-			reportLineItems.add("Class Name");
-			reportLineItems.add("Command Name");
-
-			for (String propertyName :
-					poshiProperties.testCSVReportPropertyNames) {
-
-				reportLineItems.add(propertyName);
-			}
-
-			reportCSVFileWriter.write(StringUtils.join(reportLineItems, ","));
-
-			reportLineItems.clear();
-
-			for (String testCaseNamespacedClassCommandName :
-					_testCaseNamespacedClassCommandNames) {
-
-				Matcher matcher = _namespaceClassCommandNamePattern.matcher(
-					testCaseNamespacedClassCommandName);
-
-				if (!matcher.find()) {
-					throw new RuntimeException(
-						"Invalid namespaced class command name " +
-							testCaseNamespacedClassCommandName);
-				}
-
-				reportLineItems.add(matcher.group("namespace"));
-				reportLineItems.add(matcher.group("className"));
-				reportLineItems.add(matcher.group("commandName"));
-
-				Properties properties =
-					_namespacedClassCommandNamePropertiesMap.get(
-						testCaseNamespacedClassCommandName);
-
-				for (String propertyName :
-						poshiProperties.testCSVReportPropertyNames) {
-
-					if (properties.containsKey(propertyName)) {
-						String propertyValue = properties.getProperty(
-							propertyName);
-
-						if (propertyValue.contains(",")) {
-							reportLineItems.add(
-								StringUtils.join(
-									ArrayUtils.toArray(
-										"\"", propertyValue, "\"")));
-						}
-						else {
-							reportLineItems.add(propertyValue);
-						}
-					}
-					else {
-						reportLineItems.add("");
-					}
-				}
-
-				reportCSVFileWriter.write(
-					"\n" + StringUtils.join(reportLineItems, ","));
-
-				reportLineItems.clear();
-			}
-		}
-		catch (IOException ioException) {
-			if (reportCSVFile.exists()) {
-				reportCSVFile.deleteOnExit();
-			}
-
-			throw new RuntimeException(ioException);
-		}
 	}
 
 	private static void _writeTestGeneratedProperties() throws Exception {
