+++ modules/test/poshi/poshi-runner/src/main/java/com/liferay/poshi/runner/report/PoshiReportGenerator.java
@@ -10,18 +10,32 @@ import com.liferay.poshi.core.PoshiProperties;
 import com.liferay.poshi.core.elements.ExecutePoshiElement;
 import com.liferay.poshi.core.elements.PoshiElement;
 import com.liferay.poshi.core.util.FileUtil;
+import com.liferay.poshi.core.util.StringUtil;
 import com.liferay.poshi.core.util.Validator;
 import com.liferay.poshi.runner.util.DateUtil;
 
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+
 import java.net.URL;
+
+import java.text.SimpleDateFormat;
 
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Properties;
 import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.ArrayUtils;
 
 import org.dom4j.Element;
 
@@ -35,6 +49,9 @@ public class PoshiReportGenerator {
 	public static void main(String[] args) throws Exception {
 		if (_poshiProperties.reportType.equals("usage")) {
 			_generateMacroUsageReport();
+		}
+		else if (_poshiProperties.reportType.equals("csv")) {
+			_writeTestCSVReportFile();
 		}
 	}
 
@@ -199,6 +216,95 @@ public class PoshiReportGenerator {
 		sb.append(")");
 
 		FileUtil.write(_USAGE_DATA_JAVA_SCRIPT_FILE_PATH, sb.toString());
+	}
+
+	private static void _writeTestCSVReportFile() throws Exception {
+		if (_poshiProperties.testCSVReportPropertyNames == null) {
+			return;
+		}
+
+		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("MM-dd-yyyy");
+
+		File reportCSVFile = new File(
+			StringUtil.combine(
+				"Report_", simpleDateFormat.format(new Date()), ".csv"));
+
+		try (FileWriter reportCSVFileWriter = new FileWriter(reportCSVFile)) {
+			List<String> reportLineItems = new ArrayList<>();
+
+			reportLineItems.add("Namespace");
+			reportLineItems.add("Class Name");
+			reportLineItems.add("Command Name");
+
+			for (String propertyName :
+					_poshiProperties.testCSVReportPropertyNames) {
+
+				reportLineItems.add(propertyName);
+			}
+
+			reportCSVFileWriter.write(StringUtils.join(reportLineItems, ","));
+
+			reportLineItems.clear();
+
+			for (String testCaseNamespacedClassCommandName :
+					PoshiContext.
+						getTestCaseNamespacedClassCommandNamesNames()) {
+
+				Pattern namespaceClassCommandNamePattern =
+					PoshiContext.getNamespaceClassCommandNamePattern();
+
+				Matcher matcher = namespaceClassCommandNamePattern.matcher(
+					testCaseNamespacedClassCommandName);
+
+				if (!matcher.find()) {
+					throw new RuntimeException(
+						"Invalid namespaced class command name " +
+							testCaseNamespacedClassCommandName);
+				}
+
+				reportLineItems.add(matcher.group("namespace"));
+				reportLineItems.add(matcher.group("className"));
+				reportLineItems.add(matcher.group("commandName"));
+
+				Properties properties =
+					PoshiContext.getNamespacedClassCommandNameProperties(
+						testCaseNamespacedClassCommandName);
+
+				for (String propertyName :
+						_poshiProperties.testCSVReportPropertyNames) {
+
+					if (properties.containsKey(propertyName)) {
+						String propertyValue = properties.getProperty(
+							propertyName);
+
+						if (propertyValue.contains(",")) {
+							reportLineItems.add(
+								StringUtils.join(
+									ArrayUtils.toArray(
+										"\"", propertyValue, "\"")));
+						}
+						else {
+							reportLineItems.add(propertyValue);
+						}
+					}
+					else {
+						reportLineItems.add("");
+					}
+				}
+
+				reportCSVFileWriter.write(
+					"\n" + StringUtils.join(reportLineItems, ","));
+
+				reportLineItems.clear();
+			}
+		}
+		catch (IOException ioException) {
+			if (reportCSVFile.exists()) {
+				reportCSVFile.deleteOnExit();
+			}
+
+			throw new RuntimeException(ioException);
+		}
 	}
 
 	private static final String _USAGE_DATA_JAVA_SCRIPT_FILE_PATH;
